// src/schemas/ocorrenciaSchema.ts
import { z } from 'zod';

export const ocorrenciaSchema = z.object({
  placa1: z.string(),
  placa2: z.string().optional(),
  placa3: z.string().optional(),
  modelo1: z.string().optional(),
  cor1: z.string().optional(),
  cliente: z.string(),
  tipo: z.string(),
  tipo_veiculo: z.string().optional(),
  coordenadas: z.string().optional(),
  endereco: z.string().optional(),
  bairro: z.string().optional(),
  cidade: z.string().optional(),
  estado: z.string().optional(),
  cpf_condutor: z.string().optional(),
  nome_condutor: z.string().optional(),
  transportadora: z.string().optional(),
  valor_carga: z.number().optional(),
  notas_fiscais: z.string().optional(),
  os: z.string().optional(),
  origem_bairro: z.string().optional(),
  origem_cidade: z.string().optional(),
  origem_estado: z.string().optional(),
  prestador: z.string().optional(),
  inicio: z.coerce.date().optional(),
  chegada: z.coerce.date().optional(),
  termino: z.coerce.date().optional(),
  km: z.number().optional(),
  km_inicial: z.number().optional(),
  km_final: z.number().optional(),
  despesas: z.number().optional(),
  despesas_detalhadas: z.array(z.object({ tipo: z.string(), valor: z.number() })).optional(),
  descricao: z.string().optional(),
  fotos: z.array(z.object({ url: z.string(), legenda: z.string().optional() })).optional(),
  resultado: z.string().optional(),
  status: z.string().optional(),
  encerrada_em: z.coerce.date().optional(),
  data_acionamento: z.coerce.date().optional(),
}); 